fix: add missing netlify graph tokens in env
